requirements.txt
opencv-python-headless
easyocr
numpy
pillow
python-dotenv
presidio-analyzer
spacy
google-generativeai
fastapi
uvicorn[standard]
python-multipart
